+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [unreleased]
 
+* Allow for missing .env file -- in this case the command will be executed
+  without setting any environment variables. The previous behaviour was to
+  fail with a FileNotFoundError
 * refactored __main__.py into cli.py and wrapped argparsing into dedicated
   function
 * bumped minimal Python version to 3.6